[Gold II] Title: 합이 0인 네 정수, Time: 3416 ms, Memory: 252156 KB -BaekjoonHub

+++ 백준/Gold/7453. 합이 0인 네 정수/합이 0인 네 정수.cc
@@ -29,7 +29,7 @@ int main(){
         tmp = -sumAB[i];
         auto it = lower_bound(sumCD.begin(), sumCD.end(), tmp);
         if(it == sumCD.end())continue;
-        if(*it == tmp) ans += (long long)(upper_bound(sumCD.begin(), sumCD.end(), tmp)-it);
+        if(*it == tmp) ans += (upper_bound(sumCD.begin(), sumCD.end(), tmp)-it);
     }
     cout << ans;
 }
